+++ src/bevy_ext/mod.rs
@@ -1,9 +1,11 @@
 mod cursor;
 mod fonts;
+mod picking;
 mod plugin;
 mod texture_atlases;
 
 pub use cursor::*;
 pub use fonts::*;
+pub use picking::*;
 pub(crate) use plugin::*;
 pub use texture_atlases::*;