+++ src/bevy_ext/plugin.rs
@@ -12,6 +12,7 @@ impl Plugin for BevyExtPlugin
     {
         app.add_plugins(CursorPlugin)
             .add_plugins(FontExtPlugin)
+            .add_plugins(PickingPlugin)
             .add_plugins(TextureAtlasExtPlugin);
     }
 }